multilib/bitbake.conf: default to contain MLPREFIX in RECIPE_SYSROOT
We set RECIPE_SYSROOT to ${WORKDIR}/recipe-sysroot in bitbake.conf and
${WORKDIR}/${MLPREFIX}recipe-sysroot in multilib.conf.

The problem is about nativesdk recipes. MLPREFIX is set to "nativesdk-"
in nativesdk.bbclass. So if we don't include (or require) multilib.conf,
nativesdk recipe's RECIPE_SYSROOT is 'recipe-sysroot'; if we include (or
require) multilib.conf, its RECIPE_SYSROOT is 'nativesdk-recipe-sysroot'.

Just including multilib.conf or not would affect nativesdk reicpes
RECIPE_SYSROOT value. This is not reasonable. And in fact it sometimes
causes problem for us. When we change configuration about multilib
in the same build directory, we sometimes get the following build failure.

  ld: cannot find crtn.o: No such file or directory

Not only that, sometimes the manifest under sysroot-components of
nativesdk recipes have entries starting with 'recipe-sysroot', but
the actual files that should be removed are under 'nativesdk-recipe-sysroot'.
And that causes problem for prepare_recipe_sysroot task! Errors are like:

  Exception: FileExistsError: [Errno 17] File exists

So change RECIPE_SYSROOT to contain MLPREFIX by default. In this way,
nativesdk recipes would have their RECIPE_SYSROOT set to
nativesdk-recipe-sysroot. After all, changing MULTILIBS should not
affect this value for nativesdk recipe.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>