buildtools-tarball: add nativesdk-libnss-nis
Recent glibc change removed libnss-nis module from glibc and a new
recipe libnss-nis.bb was added.

After this change, we need to make sure nativesdk-libnss-nis is also
included in buildtools-tarball, otherwise, we may encounter the following
error when using 'tar' command from buildtools-tarball.

  tar: relocation error: /lib/x86_64-linux-gnu/libnss_nis.so.2: \
       symbol _nsl_default_nss version GLIBC_PRIVATE not defined \
       in file libnsl.so.1 with link time reference

This error occured on my ubuntu16.04 host with 'nis' configured in
/etc/nssswitch.conf.

So add nativesdk-libnss-nis to buildtools-tarball to fix this problem.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>