oeqa/runtime/cases/selftest.py: rename to _selftest.py
This test modules is designed to be invoked only by selftest. It's
not meant to be tested by normal runtime test. So it should be renamed
with '_' prefix, so that it will not be automatically loaded by normal
runtime tests when 'auto' is in TEST_SUITES.

The failure message is as below.

  RESULTS - selftest.Selftest.test_install_package - Testcase -1: FAILED

Also, change selftest/cases/runtime_test.py to use '_selftest' accordingly.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>

fix _selftest

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>