dropbear: drop obsolete patch 0004-fix-2kb-keys.patch
The origins of the patch date back to early 2005 (prior to the start
of git history in oe-core) to fix a hardcoded limit on the maximum
size of remote host keys:

  http://familiar.handhelds.narkive.com/b1VGg2bI/problem-w-dropbear-ssh

The hardcoded limit was fixed upstream in dropbear 0.47:

  https://github.com/mkj/dropbear/commit/736f370dce614b717193f45d084e9e009de723ce

The patch has therefore been obsolete since then. It went unnoticed
until now as the patch has continued to apply - it modifies a value
which is not used.

(From OE-Core rev: b26cc4c4599885e8100ca751fda11417bf74b161)

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>