watchdog: make init script start after syslog
We need to make sure logs are not lost. As long as there's no
strong and vaild reason, we should make daemons start after syslog.

As a side effect, we could check the logs to see if there are
some potential problem. In OE, the 'parselogs' test case could
do it automatically.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>