waf: Fix building recipes with EXTERNALSRC defined
waf does not support being executed from another directory than the one
containing the wscript file.

Normally this is not a problem as CWD will be the same as the one
containg the waf build tool script and the wscript file. However when
building with EXTERNALSRC defined for a recipe waf would be executed
with CWD set to the normal build dir, where it cannot find wscript
and thus fails.

The solution is to change CWD to the dir containg the waf script
before executing waf.

Signed-off-by: Bjarke Freund-Hansen <bjarkefh@gmail.com>