bitbake-worker: Pass variables back to run queue
Sets up a mechanism where tasks can report information back to the
runqueue. This is accomplished by reporting back the value of any
variable listed in BB_RETVARS back from the bitbake-worker process to
the runqueue.

This obviously only works for python tasks, since they are the only ones
that can modify bitbake variables.

I'm not convinced this is necessarily the best mechanism for passing
data from tasks to the runqueue, but it allows me to keep moving forward
for now.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>