cmake.bbclass: fix wrong toolchain flags of nativesdk package
When enable multilib in an existing build project, if already build a
nativesdk package which inherits cmake, rebuild it will cause configure
error when test compiler:

| .../libexec/x86_64-pokysdk-linux/gcc/x86_64-pokysdk-linux/8.2.0/ld:
| cannot find crt1.o: No such file or directory

The STAGING_DIR_TARGET is different when multilib is enabled and
disabled. And it is a component of compiler option '--sysroot'. The
values are stored in file ${WORKDIR}/toolchain.cmake which is generated
by task generate_toolchain_file. But the task doesn't be re-run with
multilib status change.

Add ${STAGING_DIR_TARGET} to var flag 'vardeps' of generate_toolchain_file
to make sure file toolchain.cmake is re-gerated when toggle multilib.

Signed-off-by: Kai Kang <kai.kang@windriver.com>