libc-headers: update to 4.18
Updating to the 4.18 headers to match the newest kernel that will
be part of the release.

4.18 brings a requirement on bison-native to the libc-headers, since
it is required as part of the configuration steps.

We also tweak the license md5sum, since the kernel now includes SPDX
headers in the license file and that changes our sum.

(From OE-Core rev: f3b755669446872586213688ce05f579cd72dbf7)

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>