staging.bbclass: avoid staging native components for target sysroot
In the staging_populate_sysroot_dir function, components are staged
by pkgarch matching. In case of x86-64 targets, it's possible that
the native components are also matched, and thus are expectedly installed.

The negative effect is at least unnecessary installation, and in some
cases is compilation error. For example, when using eSDK and we are trying
to cross compile some project in lib32 environment, the compilation will
fail.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>