busybox: fix the alternatives logic
The current alternatives setting logic in busybox is not ideal.
For example, the 'syslogd' alternative should be handled by
busybox-syslog package, but it is now handled by the busybox package.

In other words, despite the fact that busybox recipe has multiple
packages, the alternatives settings are all done for the main busybox
package.

This will cause problems. For example, if we have an image with sysklogd
and busybox installed, there will be do_rootfs failure if busybox is
installed after sysklogd.

Note that in the above situation, the image does not have busybox-syslog.
The 'syslogd' entry should really be handled by busybox-syslog instead of
busybox.

And the same logic applies to other packages in the busybox recipe.

So change the alternatives handling codes in busybox to fix this problem.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>