os-release: fix install file conflict for multilib
It fails to create image when install os-release and lib32-os-release both:

| file /etc/os-release conflicts between attempted installs of
| os-release-1.0-r0.core2_64 and lib32-os-release-1.0-r0.x86

The /etc/os-release is a symlink link to ${libdir}/os-release. Actually
the content of files are identical and make /etc/os-release to be hard
link could fix the issue. But according to os-release (5), symlink link
is necessary for initrd environment such as dracut.

Signed-off-by: Kai Kang <kai.kang@windriver.com>