libsdl2: disable sndio explicitly to avoid inconsistent result
Explicitly disable sndio to avoid inconsistent result on different
hosts having or not having libsndio.

This must be done if we take into consideration of eSDK.

I built eSDK on Ubuntu 14.04 which has libsndio installed, and then
installed the eSDK on Fedora 27, which does not have libsndio. In
fact, on Fedora 27, I even cannot find libsndio in its repo.

The problem happens when trying to use `devtool runqemu'. The qemu
binary built on Ubuntu 14.04 needs to link to libsndio, and thus
the following error.

  runqemu - ERROR - Failed to run qemu: /path/to/qemu-system-x86_64:
            error while loading shared libraries: libsndio.so.6.1:
            cannot open shared object file: No such file or directory

So explicitly disable sndio for libsdl2 to avoid the above problem.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>