webkitgtk: fix to build without opengl
Currently, webkitgtk cannot be built if 'opengl' is not in distro
feature. The error message is like below:

  OpenGLShims.h:23:10: fatal error: GL/gl.h: No such file or directory

The required distro feature for webkitgtk is only 'x11'.

From its source codes, libepoxy is an alternative option to build webkitgtk.
So fix to use it. In this way, we can successfully build out
webkitgtk if 'opengl' is not in distro features.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>