devtool: handle do_patch postfuncs that copy files from workdir
Currently the "at" recipe in OE-Core has a postfunc on do_patch which
copies files from ${WORKDIR} to the source tree. However, if you run
"devtool modify at" (or "devtool extract at"), then during extraction
before do_patch has a chance to execute, the files it is going to copy
are moved to the local-files directory, so when the postfunc does
execute it fails. At this point the work directory is a temporary
directory so we can simply copy the files rather than moving them and
then the postfunc succeeds.

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>