devtool: disable garbage collection in new repos
When running devtool modify (or devtool extract) on some recipes (e.g.
db, gdb), since the source comes from a tarball we initialise our own
git repository, but then we move it to another directory. The latter
causes a problem because by default, automatic garbage collection is
enabled and that can be running in the background while we're moving the
repository, which causes the move code to fail as it's trying to delete
files that are already gone. This probably only happens on systems
where /tmp is on a different filesystem e.g. tmpfs. The solution is to
disable the automatic garbage collection when we set up the repository.

(I found this issue by running devtool-stress script.)

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>