devtool: fix handling gcc recipes
The gcc recipes use a shared work directory populated by a separate
gcc-source recipe, and thus they disable do_unpack and do_patch. That
caused devtool's source extraction to fail immediately because it
relies on those being there. Add some code to redirect to the
gcc-source recipe for gcc in order to make things work and allow devtool
modify/upgrade/extract to work with gcc and related recipes.

(I considered taking a copy of the work-shared directory but then
immediately ruled it out, since that would mean we would not have a
proper git tree with the patches we are applying as commits.)

Fixes [YOCTO #13036].

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>