lib/bb/checksum: fix handling filenames containing spaces/colons
There were two places in the file checksum code where we split the list
of files into items and then split the items into <filename>:<exists>
pairs. We did discover that we weren't handling spaces in filenames back
in bitbake rev 87282b283921a58426f24fb21151db457c5bca66 but we only
fixed one place. Split out the corrected code into a function and make
both places call it. Additionally, the new function was tweaked so that
it splits on the last : rather than the first, and only splits once, so
that any colons in filenames won't be split on. (If a filename happened
to have :False or :True in it we'll still have an issue, but I think
that's rather unlikely.)

This bug was triggered in OE if you used "devtool modify cmake" followed
by "bitbake cmake".

Signed-off-by: Paul Eggleton <paul.eggleton@linux.intel.com>