staging.bbclass: Avoid looking at other multiconfigs on extend_recipe_sysroot
On extend_recipe_sysroot we only compare taskname and pn, but if we have
some dependency to a recipe from another multiconfig, these two will
still match if it is the same recipe, causing an error since the build
system believes it has already proccesed that dependency.

Check if this is a multiconfig build, if it is, make sure the current
dependency is coming from the same multiconfig as the one we are
building.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Ross Burton <ross.burton@intel.com>