lib/oe/patch.py: Fix applying a directory as a patch
If a SRC_URI content ends with '.patch' bitbake is
trying to apply it as it's a patch file.

It causes that if we use git repository for 'patch' package
the bare clone is extracted to a directory
(i.e. build/downloads/git2/git.mirror.org.patch/) which is considered
to be a patch file, so patch.py tries to apply that directory as a patch
which ends up with a failure.

Signed-off-by: Tomasz Dziendzielski <tomasz.dziendzielski@gmail.com>